Fluid dynamics symbolic generation
Renaming the documentation folder to symbolic_generation. Update of the NavierStokes symbolic element to use Picard linearisation by default.  The documentation of this element has been also updated (note that this will be moved to its proper place once the documentation discussion is finished). Slight correction of the manufactured solution test expected results due to the element modification.