Fixing problm with new schemes
The scheme needs to give a pointer to itself to the Properties now. Since no function had been provided to the hybrid-bashforth scheme it did not work. Now the no-history-force case works fine again. The history force case gives large errors in the Candelier example. I have looked through many of the files changes since it did and found nothing...