Apply outlet process update
Now the outlet process uses the "experimental" processes in the core for setting both the PRESSURE (set and fixed) and the EXTERNAL_PRESSURE (just set) variables. Besides, the flag OUTLET is set in both the outlet model part nodes and conditions.  
Finally, the average outlet velocity is computed and stored in the CHARACTERISTIC_VELOCITY variable (this value is intended to be u sed in the outlet inflow prevention condition recently implemented).