Fixing periodic domain contact
I added new steps that werer ignored in spheric_particle and created a bins specialized to periodic conditions. Now aparently they work with a colloid-extrusion example (only perhaps very slight boundary effects are still visible, although this could just be a confusion). The fix is not ideal in that it does not include the default, non-periodic case and ifs will need to be added. i also did some minor optimizations here and there.