making the ethier_becnhmark work with latest script form
Now the same slopes are obtained in the order-of-convergence slopes. The set of cases are launched from the run_ethier_benchmark.py file, unlike before (we used a bash file). There still apear errors at the end of each run, but they do not affect the results. It is probably something to do with memory leaks for not destroying some objects in python. This will be investigated next