BeginImplementing J2 SS
Begin implementing J2 for SS. For the moment, no changes are introduced in non_linear_assosiatvei... since the same procedures may be used to compute the CauchyStressTensor (the TangentMatrix, Update StrainTensor are not yet included). Further changes are needed to compute the trial state (for the moment I begin from everything zero. Do not FinalizeSolutionStep since will create strange behaviours.