Avoiding adding so many varaibles and adding check for them
I have modified the criteria to add some of the varaibles, since it included by default too manyu variables. In case anything breaks, we can have a look, but this was necessary. I have also added a nice function to check if the model part has that nodal varaible from python