Adding tool to track history of all particles
It is useful to have a way to track the particles that have existed during a simulation in a robust way. For example making statistical inferences. Creator distructor takes an optional watcher that by default does nothing. This keeps all the info in std::vectors. There now needs to be a way to clear the data periodically. Still only a prototype. Next, implementation.