Making the particle history watcher work
Now it records all particles that ever existed in a hdf5 file. Also adding a tool to plot the probability of ending inside a bounding box, which is implemented in hdf5_tools.py. This file has to be reorganized and separate the fluid loader from the particles loader. This has just been started by creating the ParticleHistoryLoader. I have to think about a central object to concentrate all the tools for hdf5.