Simplifying, making modular and making safer
Now the possibility of being dettached with no neighbours involves no security problems. THe only thing is that in the unlikely case that the particle is dettached with no neighbours (search radius must be very tight), there will be a one-step innaccuracy in rotating the inlet velocity. iN my opinion such error is negligible.