Reservoir temperature condition fixed
In the older version, the temperature condition for the dry part of the dam, was assigned based on an 'Outer temperature parameter': impose temperature condition. Now, this statemen has been removed and the dry part of the upstream face of the dam remains the condition imposed previously for the whole dam (upstream and downstream). Thus, it is coherent with the rest of the dry faces (i.e. the same temperature condition or flux imposed).