Impose 'Nodal Reference temperature' for thermal_linear processes
A new variable was created through the DamInterface and transfered to Kratos: 'NODAL_REFERENCE_TEMPERATURE'. This variable assigns to each node a constant value of 'reference temperature' or a different one, the last option importing them from an external file.
One the variable is created and introduced in its corresponding process, all the constituve laws refer to this new variable 'NODAL_REFERENCE_TEMPERATURE' to calculate the 'ThermalStrain' Vector.