Completing the transition to analysis stage in SDEM
Now SDEM has the overall form of the analysis_stage. This has required creating a SDEM solver, to which many of the operations previously realized in the stage are now done there. The overall scheme of the application is as follows: SDEM_stage contains a solver (the new SDEM solver), two stages (the modified DEM and FluidDynamics stages) with their corresponding solvers. This scheme is not completely general for coupling applications, but it seems that it might do here. All DEM tests and the Candelier test pass, except the occasional random error in one of the DEM tests, which will be studied next.