Adding history force law and a new test
Unfortunately, this is an intermediate stage at which the candelier tests run fine, except for the last one that uses the MAE. If the function ComputeBassetForce from the swimming_sphere is called as well as the one from the hydrodynamic_law (it should be redundant), all tests run ok. It is left to find the error.