refactor(HDF5): restructured python classes in hdf5
The HDF5Application.core python classes are organized into:
- processes
- controllers
- IOs
- operations

The method HDF5Application.core.Factory(settings, model) can be used to
construct a custom HDF5 IO process specified by the json settings and serves as
the back end to the simpler but less general process modules.

The new code also supports the non-terminal symbols <identifier> and <time>,
which allow the model part name and time to be expressed in either file path
strings or file prefixes.

Unit tests are included.