Removing old files, getting rid of pp object and some cleanup
The most important change is the removal of the algorithms, which nobody is using anymore, the elimination of the pp object that caused a lot of confusion and some cleanup. The Chandelier tests run and a simple example that includes interpolation does too. Next, creating a test including an inlet and recovery.