Eliminating UpdatedFluid/UpdatedDEM distinction
This distinction is not relevant anymore, so I have decided to remove it. The fluid is not by default ahead of the DEM, which must catch up with it. This could not be optimal in two-way coupling cases, but I think it is better to treat that case separately and a new solver should perhaps be created for such a case.