Replacing SLIP_VELOCITY by AUX_VELOCITY and correcting small bug
The alpha parameter was not being passed to the function in charge of evaluating the intermediate AUX_VEL (before SLIP_VELOCITY) at the particles' locations, which is not consistent with the Candelier test (and the original algorithm surely) . All tests work except for the PFEM test (it complains about a variable not being added in PFEMApplication). After setting 'bounding_box["use_bounding_box"].GetBool()' to False manually it works too.