fix(StructuralMechanics): fixed kirchhoff plane stress calculation
The kirchhoff plane stress law was computing PK2 stresses incorrectly from the
base class implementation. This bug caused incorrect stresses to be computed
when the flag COMPUTE_CONSTITUTIVE_TENSOR was set to false.

This resulted in the correct stresses in BuildAndSolve() but incorrect stresses
in BuildRHS() for the TotalLagrangian element, which caused the convergence
criteria to behave incorrectly.