feat(StructuralMechanics): added total lagrangian elements to adjoint solver
The changes add total lagrangian adjoint elements to the list of available
adjoint elements in the structural_mechanics_adjoint_static_solver. The elements
are already tested in the cpp unit tests. In order to interface with the
replacement process, the second Create() method needed to be added to the adjoint
element. Also, to avoid checking for primal dofs, the Check() method was
modified.