refactor(HDF5): added docstrings and parameters wrapper
The docstrings provide a summary of the json settings for configuring each
HDF5 IO process. The parameters wrapper was added to improve readability
when mapping user parameters to core parameters while keeping the amount of
code about the same. Minor renaming of classes and variables was also done.

The behavior of each process is not modified. The frequency of Xdmf output in
the HDF5-Xdmf output process was adjusted to only write metadata if HDF5 data
was written. Previously it was writing at every step.