[Core][Proposal][NOT  TO MERGE] Adding create to Process
This is added in order to have a discussion. As @rubenzorrilla stated some time ago, it is inconsistent and duplicative to create a C++ process to later create a python scripts that just calls that very sma process. In order to avoid that we need to have a kind of factory, but I think it can be achieved in a simpler manner simply adding a Create method using the same inputs as in python (Model+Parameters) in  first phase. In a second phase we should add the Processes to the KratosComponents database so wen can create them just with the name and add it to the process factory