fix(HDF5): fixed issues with tests
After #5437, failing test were reported on some systems. These appear to
be caused by a clash between the HDF5 linked to the c++ tests and the
HDF5 used by h5py. We resolve the problem here by running the c++ tests
in a subprocess.

A test was occasionally failing due to an invalid assumption that the
constructed list of XDMF nodal results is partially ordered by the type
of nodal container. This splits the test to eliminate the bug.
Alternatively an ordered dictionary may be used in the implementation of
XdmfNodalResults().