refactor(RANS): making test and element names more consistent
Unit test file names are changed to make it more obvious which files
they test. The original file name is prepended with `test_`. For example,
unit tests for evm_k_adjoint_element.h are in
test_evm_k_adjoint_element.cpp.

Tests are renamed to follow a common naming convention. This makes it
easier for the reader to figure out which member function is tested.
For example, a unit test for
`EvmEpsilonAdjointElement<2, 3>::CalculateLeftHandSide(..)` is named
`EvmEpsilonAdjointElement2D3N_CalculateLeftHandSide`. To differentiate
between overloads, one can append an additional underscore and keyword to
to test name.

Some long string names in the tests which have no effect are shortened
and the element names for registering in kratos components are also
made more consistent.