refactor(RANS): element interface migration
The configuration of the monolithic element matrix construction through the
process info is removed from the VMSAdjointElement and the template parameters
are simplified. The interfaces of VMSAdjointElement, EvmKAdjointElement and
EvmEpsilonAdjointElement are extended to include overloads based on bounded
matrices and vectors. These are to be used by the monolithic element instead
of the regular matrices with monolithic construction. They are also used
internally for the implementation of the regular Element methods. This avoids
unnecessary heap allocation in the monolithic element routines while keeping
the original behaviour and simplicity of the VMSAdjointElement interface.
EvmKAdjointElement and EvmEpsilonAdjointElement members still have monolithic
construction dependencies, but these are no longer required by the monolithic
element and can be removed. The Calculate() method of the EvmKAdjointElement
and EvmEpsilonAdjointElement can most likely also be removed.