[Core] Adding new test looking for error in #8096
Trying to find the error from #8096. I added a new test that works in OMP. I do not  know how to trigger the problem. In theory it appeared when the dof was only in the MPC but not in the element. This small test does that but does not trigger the problem, any idea why?