update element base interface [ unit tests are changed ]
Unit tests were changed because, the tests are 2D,
But all three components of vectors are assigned and norm_2
is used to get the norm. This PR ports to use array_1d<double, TDim>
so then the norm is different [ it ignores the third component in 2D
case. The changes were not computed in this branch. It was computed
in the master by making the third component zero in the 3D vectors and
the changed values from that master is transferred here. ]