Minor error in the exp part of the equiv ep calculation
Error in the formula of the equivalent plastic strain calculation. Now obtained results seem to be good; still needed a check with excel file (but the relation ep kp is still needed for that and so will take long). Despite this, the obtained curve is smooth and no error is expected to be found here