Computing Sult in a new method + adding condition in Nlocal calculator
Computing the Sult in a new method is required as the unified fatigue CL introduces the Sult as an input. The new condition of the Nlocal calculator is (1-dred) * Smax >= sth is added because B0 is not recomputed and in the case that it was computed Nlocal should be inf and can induce an error.