LSPG ROM Application Least Squares Petrov Galerkin is a projection approach that minimizes the residual in the least squares norm at each iteration.
Yielding a linearized system of equations at each nonlinear operation of the form:
(LHS)Phi (dxrom) = RHS

Where the LHS contributions for the fixed dofs are not considered.

This branch is intended to be merged into the master branch in order to investigate the possibilities of developing a Petrov-Galerkin technique and to learn further about it's compatibility with Neural Networks. This technique was originally explored in PR #9735, but it was determined to be divided into pieces, the first of which is this one.