Use `= default` instead of an empty functionbody for destructors
Destructors are special member functions that are overridden when the
base class destructor is `virtual`. Don't supply empty function bodies,
but use `= default` instead. Note that this is in line with C.80: "Use
`=default` if you have to be explicit about using the default semantics"
from the C++ Core Guidelines.