No longer create a symlink to the compilation database
Creating a symlink on Windows requires elevated rights. As a result, the
following error was raised by CMake when a regular user attempted to
build Kratos on a Windows platform:
```
CMake Error: failed to create symbolic link
'C:/path/to/kratos/source/compile_commands.json': A required privilege
is not held by the client.
```
This error was triggered by an attempt to create a symlink in the source
directory that points to file `compile_commands.json` in the build
directory.

In consultation with the original author of the CMake code involved
(Máté Kelemen) we have decided to remove the command that attempts to
create the symlink, since it is no longer needed.