Fixed a slicing issue
Avoid constructing a (temporary) object, which slices off a part of the
original object. Use a reference-to-const instead. Take into account
Phil Bucher's comment about not using `auto` for the reference type. The
problem he pointed at may be related to the lazy evaluation of
expressions by uBLAS. To be investigated further.