Working in parellel. Read comments.
It is working in parellel, nonethelesss, this version is  building the projected residuals with OMP instead of parallel utilities because of the ordering of the elements (k). It is also worth noting that this  model is doing some tricks. The dynamic LROB  is constructed with all model part. And please notice I did some changes in the size of the system, I am checking what happens for HROM cases (the EquationSystemSize).