Reduce c and phi attained from the model part's properties
The model part keeps its own copy of the properties (which should not
change over time). Use those as the starting point when applying the
reduction scheme for c and phi.

Previously, we retrieved the (reduced) properties kept by the elements
and reduced those. In that way, the reduction went much quicker than
anticipated. Also, we had no way to access the original properties. Now
we do.