Added a process factory to be used by custom workflows (in C++)
The factory first needs to be populated with creators for concrete
process types that need to be produced. On request, it will produce a
smart pointer to a `Process` instance, thereby supplying a uniform and
consistent interface to the processes.

This commit includes the implementation of the factory and a set of unit
tests. The actual usage by a custom workflow will be part of another
commit.