Added an adaptive time incrementor class
This class is intended to be used in a time stepping scheme, where the
size of the time increment may vary depending on the convergence rate.
In other words, when convergence is achieved in just a few iterations,
the time increment for the next step may grow. If, on the other hand, no
convergence is found, it allows to retry the time step with a smaller
time increment.

Note that this commit also includes a time incrementor interface, which
allows us to implement different kinds of time incrementors (e.g. a
collection of prescribed time increments).