Moved logic of `ApplyScalarConstraintsTableProcess` to C++
This class used to be defined in a Python module only. Now that we also
need it for the custom settlement workflow, we had to port it from
Python to C++. Note that the Python interface now uses the new C++
class.

Also added a creator for the `ApplyScalarConstraintsTableProcess` to the
process factory used by the custom workflow. The factory should now be
able to create all kinds of processes that are relevant for settlement
analysis.