[GeoMechanicsApplication] Add support for non-uniform line loads (#11658)
Added support for applying non-uniform line loads. It does so by attaching a table to a uniform line load, where the table defines the relationship between a set of X coordinates and the corresponding "pressure" values. The pressures applied to nodes will be linearly interpolated (and/or linearly extrapolated, if needed) using this table.