Extracted copying of nodal solution step values
When the solving strategy wrapper needs to restore the nodal positions
and the nodal solution step values, it explicitly checks for each nodal
solution step variable whether the model part actually has it. If yes,
then it copies for each node the value at the given source index to the
given destination index.

Split an existing unit test into two new tests. The first one verifies
whether the combination of displacements and water pressures can be
restored. The second one verifies whether the combination of
displacements and rotations can be restored.