Create a computational model part
In line with the approach followed by the Python scripts, we need to
have a computational (sub)model part that keeps "active" parts of the
mesh/model, including the nodes, elements, and conditions of the main
model part. It also shares the process info with the main model part.
This computational model part is then passed to the factory that creates
solving strategies.

When the solving strategy wrapper attempts to clone a time step, we now
either pass its model part (if it is not a submodel part) or the root
model part of its (sub)model part.

Furthermore, we added a few more print statements for debugging.