Simplified the writing of integration point (and nodal) results
We no longer need to cast and test a variable to determine which
function overload we need for writing the variable's results. We now
build up a map of integration point (and nodal) result writers, which in
turn are created by two factory function templates. Adding a new result
to be output is a matter of adding a single line.