[GeoMechanicsApplication] Add new element class for heat (#11753)
Added a class template for (transient) thermal elements to be used in geothermal analyses. It has been based on the work done by @mnabideltares. Afterwards, the code has been cleaned up, refactored, and restructured by @markelov208 and @avdg81.

This commit is part of a series of commits to bring in basic thermal functionality in the GeoMechanicsApplication.