Extracted a loop condition for finding the critical head
One of the loop conditions is whether the current head does not exceed
the maximum critical head yet. That logic has been moved to a separate
member function, which makes it easier to understand how the loop works.

As a result, data member `mExitLoop` was removed, since it had become
redundant.