Creating a condition for testing now requires the dimension size
To make the function that creates a micro-climate condition for testing
more generally applicable, it now has a parameter for passing the
dimension size.

Also added a unit test to verify that when initializing the solution
step of a micro-climate condition acting on a linear triangle executes
without raising any errors.