Added a test for a condition acting on a quadratic triangle
The unit test verifies that when initializing the solution step of a
micro-climate condition acting on a quadratic triangle executes without
raising any errors.

The helper function that creates the nodes for a triangle has been
extended such that it can also create the mid-side nodes of a quadratic
triangle.