Modified the unit test that verifies the RHS vector calculation
Use a quadratic line condition rather than a linear one. The idea to is
have a test for each shape of condition (i.e. line, triangle, and
quadrilateral). At the same time, we would like to limit the tests to
quadratic shape functions only (to avoid an explosion of tests).