Remove the `runkratos` executable (#11898)
This commit extracts the changes related to the removal of `runkratos` from PR #11506. In essence, this commit removes the source code of the `runkratos` executable as well as its build instructions. Furthermore, all scripts that used to call `runkratos` have been updated. Also the documentation has been updated, such that it no longer references the `runkratos` executable.