Added a unit test for the apply K0 procedure
The new test verifies whether coupling terms are considered when the new
flag `force_plaxis_compatibility` has not been supplied.

To make this work, we have added a data member to class
`ApplyK0ProcedureProcess` which keeps the settings of the K0 procedure.

Also added a stub class for a linear elastic law.