Renamed the flag to use the "standard" K0 procedure
It used to be named "force_decoupled_elasticity". Now the flag is named
"use_standard_procedure".

Also renamed the function that retrieves the flag's value.

Note that the behavior when the flag is not defined has been changed. In
that case, we now use the "standard" procedure, i.e. consider the
diagonal entries only and set the shear terms to zero.