Removed a near clone of element utilities
Apparently, in the past the custom element utilities of the
StructuralMechanicsApplication had been copied to the
GeoMechanicsApplication. To eliminate this duplication, we have removed
the (near) copy and replaced its usages by calling the corresponding
functions from the StructuralMechanicsApplication. To make this work, we
had to export two utility functions.

Note that there were some minor differences which we deemed
insignificant:
- The StructuralMechanicsApplication uses nonmember functions which are
  encapsulated in a dedicated namespace, whereas the
  GeoMechanicsApplication used a class with static member functions.
- When calculating the Rayleigh damping matrix, the
  StructuralMechanicsApplication takes into account a tiny tolerance for
  the 'alpha' and 'beta' coefficients, whereas the
  GeoMechanicsApplication accepted any positive value for them.