Refactored the unit test for `ExtractSolutionStepValues`
Added a few local helper functions to simplify the test code.

Also modified the API for extracting the DOFs for the given collection
of nodes. Now there are two function templates. The first one receives a
reference-to-const to a range of nodes. The second one receives a pair
of iterators for passing the range of nodes.