[GeoMechanicsApplication] Refactor processing of DOFs by elements and conditions (#12090)
This commit eliminates several repetitions of the order of degrees of freedom (in elements and conditions). Ideally, the order should be defined just once. This will improve the maintenance of the corresponding code. Also, it will become much easier to change the order as needed.

As a bonus, due to removing the duplications, we have also addressed a significant amount of code smells.