Extended the utilty function for creating a test model part
The utility function that creates a model part with a single 2D3N
element now receives an optional vector of nodal variables. These are
added to the model part as well as to the new nodes.

This utility function is now used by one of the unit tests for the DOF
utilities.

Also renamed a local variable in the unit test.