Added a unit test for extracting U-Pw DOFs of a 3D element
The test extracts all U-Pw DOFs of a tetra element. So far, it only
checks the size of the returned vector of DOFs.

Note that the interface of the DOF utility function has been extended:
it now also requires the model dimension. When we are dealing with a 3D
model, an extra displacement DOF is added per node. However, note that I
have added `DISPLACEMENT_X` instead of `DISPLACEMENT_Z`, to make a
future check fail.