Extended the utility function for extracting U-Pw DOFs
The aim of the extension is to support 3D diff order elements in
addition to 2D ones.

Also added a test utility function that creates a 3D diff order tetra.

So far, the new unit test only checks the number of DOFs that is
returned.

Note that the DOF pointer for the third dimension for now is equal to
`nullptr`.