Fixed a bug where the maximum number of iterations was not properly read
In the C++ workflow for settlement analyses, the maximum number of
iterations was read from setting "max_iteration", whereas we normally
use "max_iterations" (mind the extra 's' at the end of the latter). Now
the value is read from the expected setting (i.e. "max_iterations").

Note that a similar approach was already followed for the Python route.