Inlined member functions `SetRetentionParameters`
These member functions only set the fluid pressure (taken from the
`ElementVariables` data structure) that is kept by the retention law
parameters. This simple action does not need another level of
indirection.

Furthermore, it was found that in one location the fluid pressure was
calculated, but its result was not stored. Now it is.

Also marked a few member functions `[[nodiscard]]`, to trigger warnings
when the returned result is discarded.

Finally, a few more member functions have been made `const`.