Fixed the calculation of the table's derivative outside its domain
When the derivative of a table was requested at a point outside of the
table's domain, it always returned 0. This was inconsistent with the
behavior of member `GetValue`, which extrapolates using either the first
line segment or the last line segment. The calculation of the table's
derivative has been made consistent now. The change is covered by a unit
test.