[CoSimulationApplication] Reverting `acusolve_wrapper.py` to its debug phase
We have noticed that the version `acusolve_wrapper.py` is conflictive with the developments in the AcuSolve team and therefore as the code is not ours we have decided to revert to its debugging version which apparently works. This means that the style is reverted and it even includes code reverted. We may add style again once we have some way to run tests, like for example mockups solvers.