Fixed occasional crash when constructing `LineInterfaceGeometry` instances (#12679)
When a `LineInterfaceGeometry` instance was constructed with a list of node pointers where at least one of them is null, the application would crash. The underlying reason was that calculating the points of the mid-line geometry failed, due to dereferencing the null pointer. This fix checks whether there are any null pointers in the provided vector of node pointers, and if there are it will return a vector of null pointers for the mid-line points.

Note that being able to deal with such input is required, since at element registration time we always provide the "blueprint" element with a list of null pointers for the element's nodes.