Re-enable settlement test using the C++ route (#12877)
Converted a _disabled_ C++ unit test to a Python test. It shouldn't be a C++ unit test, since it tests a complete workflow, not just a single component. The new Python test covers the so-called "C++ route" of the settlement workflow. In other words, it uses the exact same entry point into the code as D-GeoSettlement does.

Since the `KratosGeoSettlement` class owns a local `KratosGeoMechanicsApplication` instance, running this new test will deregister all GeoMechanicsApplication components (e.g. geo elements) when that object is destroyed. To avoid that any subsequent tests start failing, we need to reload the relevant Kratos applications at the end of the new test (in its `tearDown` method).

With these changes, there are now two Python tests that simulate the same problem (a settlement problem), but using different entry points in our code (i.e. the "C++ route" and the "Python route"). Both ways should yield the same results. That is asserted by these tests.