[GeoMechanicsApplication] Remove number of umat state variables (#12968)
The number of umat/udsm state variables can be deducted from the length of the vector supplied by the user. As this is user input, having an output item for it seems overkill. Removed all use of NUMBER_OF_UMAT_STATE_VARIABLES, the setup of the variable and the output item.