[GeoMechanicsApplication] Removed initialization of material response from `InitializeSolutionStep` (#13027)
In [PR #13014](https://github.com/KratosMultiphysics/Kratos/pull/13014), initialization of material response has been added to the element initialization. Previously, that was done by the initialization of the solution step. It was, however, not removed there. Now it has been.

Data member `mIsInitialised` was no longer needed, since the implicit assumption of class `UPwBaseElement` is that member function `Initialize` must be called prior to calling `InitializeSolutionStep`. A handful of unit tests lacked the call to `Initialize` prior to calling `InitializeSolutionStep`, and therefore started to fail. Those tests have been fixed as well.