Use member `reserve` of `std::vector`
Rather than creating an instance of `std::vector` with the desired size (which will default-construct the vector's elements), we prefer to default-construct it and use `reserve`. That allocates the required amount of memory, but the vector itself remains empty. With this pattern, we avoid default-constructing the vector's elements that will be replaced not much later.