Added `load` and `save` members to the stress state policies
To make this work, we had to make a minor change to the (de)serialization code of Kratos Core. We've added a compile-time check to see whether a given class is an abstract class. If it is, we shouldn't try to instantiate it. In practice, this is not expected to ever happen.