Improved (de)serialization of `Kratos::intrusive_ptr`
When the data type of the intrusive pointer is an abstract base class, no longer attempt to instantiate it (which yields a compile-time error). This would only be attempted when the pointer type is `SP_BASE_CLASS_POINTER`. In practice, the underlying type will always be different from that abstract base class, and therefore the pointer type must always be `SP_DERIVED_CLASS_POINTER`, which looks up the underlying concrete type to instantiate.

To make this work with the test classes, a few more members had to be defined.