Improved (de)serialization of raw owning pointers
When the data type of the raw owning pointer is an abstract base class, no longer attempt to instantiate it (which yields a compile-time error). This would only be attempted when the pointer type is `SP_BASE_CLASS_POINTER`. In practice, the underlying type will always be different from that abstract base class, and therefore the pointer type must always be `SP_DERIVED_CLASS_POINTER`, which looks up the underlying concrete type to instantiate.

Also changed the test suite of the new unit tests to the one without a kernel, since it's not needed.