Instances can now be loaded through an abstract interface
When loading a serialized smart pointer (or an owning raw pointer), don't attempt to instantiate the corresponding data type when it is abstract. This is useful when passing a (smart) pointer to an interface class, which was not supported so far. Note that without the added check, a compile-time error would be raised when passing a (smart) pointer to an abstract class.

Other changes include:
- Added a member function to class `Serializer` to deregister data types for (de)serialization.
- Added four unit tests that verify whether an instance can be saved and loaded through a (smart) pointer to an interface class.