[Core] Support serialization through (smart) pointers to abstract classes (#13153)
This commit adds a few compile-time checks that avoid attempting to instantiate abstract base classes while loading objects from a `Serializer` instance. So far, it was not possible to use (smart) pointers to abstract base classes when serializing objects of derived classes. By adding a few compile-time checks, we now can supply (smart) pointers to abstract base classes to the existing (de)serialization code.

For each modified `load` function, a corresponding unit test has been created.

To keep the scope of changes to the registry as local as possible during a test run, a test class is registered and deregistered as needed using the RAII idiom. To this end, class `Serializer` has been extended with a member function to deregister data types.