Made some `static` members free functions
This is in line with C++ Core Guideline C.4: Make a function a member only if it needs direct access to the representation of a class. The free functions have been put in an unnamed namespace to comply with C++ Core Guideline SF.22: Use an unnamed (anonymous) namespace for all internal/non-exported entities.