[CoSim] Rigid Body Solver (#9419)
A simple solver to represent rigid bodies that move or rotate but do not deform (e.g. a bridge cross section moving inside the fluid domain or a TMD attached to a structure).

It solves each degree of freedom (3 displacements and 3 rotations) as an independent SDoF solver, but it organizes the data so it can be easily coupled which other solvers when more than one DoF is necessary. One can assign structural properties, a root point excitation or a forced motion to each of the DoFs. Each of them can also be "activated"/"deactivated", so an SDoF case can be also solved with this solver.

Additions:
  - The rigid body solver (rigid_body_solver.py, rigid_body_process.py, input_check.py)
  - A wrapper so it can be used with CoSimulationApplication (rigid_body_wrapper.py)
  - Two coupling operations to communicate with FluidDynamicsApplication if necessary (compute_resultants.py, impose_mesh_displacement.py).
  - Two data transfer operators to sum to convert from several data values to one value and vice versa (copy_single_to_distributed_vectorial.py, sum_distributed_to_single_vectorial.py).
  - Corresponding tests.