Improved handling of unit tests that need to run conditionally
Use `GTEST_SKIP` rather than not compiling unit tests that need to be run conditionally. So far, several unit tests are skipped when non-debug builds are used.

With these changes, the total number of unit tests that is run is identical for all build types. However, the number of skipped tests may differ.