Fetching nodal results from GiD output no longer requires sorted node IDs
Internally, the function that fetches nodal results for the given time now uses a map from node ID to requested value. As such, it is no longer needed to pass sorted lists of node IDs.

Also added an assertion to test that the number of fetched nodal tensors matches the number of node IDs. Currently, this assertion fails for at least one test, which indeed appeared to be an error.