Changed the material models of clay and sand to linear elastic
This is needed for the first stage only (K0 procedure). The solid densities as well as the porosity values are now in agreement with the weights that are listed in the Plaxis tutorial.

Furthermore, the retention laws for clay have been changed to "SaturatedBelowPhreaticLevelLaw", since the phreatic line divides the clay in a saturated part and an unsaturated one.