Added function objects to describe the properties of the Coulomb yield surface (#13910)
Extended the Coulomb yield surface with function objects that calculate the values of the friction angle $`\phi`$, cohesion $`c`$, and dilatancy angle $`\psi`$ given parameter $`\kappa`$. So far, we have only implemented the case where these functions always return the same value, regardless of the supplied $`\kappa`$. In fact, at this point in time $`\kappa`$ is set to zero and not updated at all. This preserves the behavior that we had before (where $`\phi`$, $`c`$ and $`\psi`$ would remain constant). Soon, we will move to formulations where we actually have a dependency on $`\kappa`$.

The primary goal of this PR is to lay the groundwork for a mechanism that we can easily extend to include more complex dependencies on $`\kappa`$ without needing to modify the overarching framework.